Working on getting num-derive decode methods for all enumerations
